Separate IndependentSet graph in separate file
